refactor: rename cli.py to create_github_app.py
Rename files for clarity:
- cli.py -> create_github_app.py
- test_cli.py -> test_create_github_app.py
